Use Python 2.3 for selecting-python tests and only run these tests if the version of python used to run the tests is not Python-2.3.
Use pypi as the index when running these tests. This is,
unfortunately, necessary because when installing a source distro for
another version of Python, we need to download setuptools for the
desired version.  Maybe in the future we can figure out a way for the
tests to use a local copy of setuptools.


git-svn-id: http://svn.zope.org/repos/main/zc.buildout/trunk@70460 62d5b8a3-27da-0310-9561-8e5933582275