Return copies of cached envs for better isolation.
Copying an instantiated environment doesn't rescan the filesystem for
dists and allows for the consumer of the env to modify it without
affecting subseqent uses of the cache.

The env cache still causes test failures.  In
zc/buildout/debugging.txt, for example, it can't find a part recipe
because the develop egg path isn't being added to sys.path while it
does get added if the cache isn't used.