Using distribute's way of chmod +x.
os.umask is a bit weird in that it *does* return the current umask,
but it has a required argument that it immediately sets as the new umask...
Weird. Distribute solves that by assigning a dummy value and re-setting
the umask to the original value.

The matching test verifies that it actually works.

Note that I use python 2.6+'s octal notation: 0o755.
Buildout 2 doesn't support python 2.5 anymore, iirc, so that's fine.