Warn when using an outdated bootstrap that gives you setuptools
Fixes #82

Thanks Marius Gedminas for the fix I could copy/paste :-)

Note that there is no test for this as the test setup doesn't allow for it.
The error that this commit fixes, fixes an error that occurs at
easy_install.py import time, so the fix also has to be on that level, making it
hard to test.
I personally think it's ok like this... ;-)