Detecting if the installer is being run to install buildout itself.
A working_set passed into install() is the sign that we're called from
buildout.py to install buildout, setuptools, an extension or a recipe.
Those need the global pkg_resources.working_set, for instance to load entry
points.

So... conflicts with global packages are possible, but they should be ignored
so that we can grab the right versions ourselves. The paths that end up in the
bin scripts that we generate are OK.