Properly handling non-ascii filenames in hashing function
Added fs_to_text() function that ensures a filename is always unicode (python
2) or str (python 3).

os.walk() gets such a unicode dirname as start point, so that it tries to do
its best on python2. The resulting filenames are converted to unicode to make
really sure that the join() doesn't fail with a decoding error.

The input to the hash function is encoded, as it gives an ascii encoding error
otherwise on python 2.