Problem: `setuptools.easy_install` is deprecated
Solution: Use `pip install`

The meat of the change is in src/zc/buildout/easy_install.py

Some tests had to be refactored a bit to make them more robust.

`tox` is removed to avoid issues due to sharing the same directory
between different virtual environments; `Makefile` enriched to make
per python virtual envs depend on source files.

No need to be afraid of default Python install as we depend on `pip`.
`bootstrap.py` is gone as not needed anymore.