Fix bootstrapping for development
- fix bootstrapping for Python 2.7

The bootstrap process uses `get-pip.py`, which in its latest version,
tries to bootstrap a Python 3 only version of pip.

We need to pass in a separate URL for a Python 2.7 compatible version of
get-pip.py.

- Handle different pip messages

e.g. on pip 20.3.4 the message for no updates is
"Requirement already up-to-date"

e.g. on pip 21.0.1 the message for no updates is
"Requirement already satisfied"