Make compatible with pip 22.2+, restoring Requires-Python functionality there.
Fixes https://github.com/buildout/buildout/issues/613.

Note: we are patching `process_url` from `setuptools`.
The existing comment says that this method was copied over from setuptools 46.1.3.
I was wondering, so I checked: the method is still the same in latest setuptools.  And it is largely unchanged since setuptools 42.0.2.
So for that part we should still be compatible with quite a long range of setuptools versions.