Revert "Temporary: completely remove the GitHub actions."
This reverts commit e63052e0a1aeaa5820b581c46ba65196d46cec80.

The first batch of CircleCI problems are fixed with the current commits.
See https://app.circleci.com/pipelines/github/buildout/buildout/334/workflows/6baea55f-e2e8-4a2a-998d-3ecc6ed64cf1
This means on Ubuntu it now works.
But now that those pass, other checks kick in, for Debian, CentOS and Mac.
They apparently need the `.github` directory, so let's re-add it.

I wonder if there is much value in testing this on CircleCI when we seem to do the same on GitHub Actions.
And I expect the CentOs tests to fail; I already disabled them on GitHub Actions.