Add config option: optional-extends.
This is the same as the `extends` option, but then for optional files.
The names must be file paths, not URLs.  If the path does not exist,  it is silently ignored.

This is useful for optionally loading a `local.cfg` or `custom.cfg` with options specific for the developer or the server.

I have seen various ways of having local options without checking this into version control:

* Let git ignore `buildout.cfg` and have a sample `buildout.cfg.in` that you copy to `buildout.cfg` and adapt for the machine.
* Have a `local.cfg` that extends `buildout.cfg` or an alternative like `devel.cfg` or `production.cfg`.
* Let `buildout.cfg` extend `custom.cfg` which is not in version control, so you get an error when you don't yet have this file.

When this PR is merged, I intend to use this in the Plone core development buildout.