Add test-recipe buildout part and 'make test-recipe'.
This runs the `zc.recipe.egg` tests.
There was olready `bin/oltest` which runs tests for which it is fine to be offline.  This includes (some of the) `zc.recipe.egg` tests.
But in both cases several test failures like the following arise:

```
Error: We don't have a distribution for demoneeded
    and can't install one in offline (no-install) mode.
```

This suggests that this includes tests that actually *do* need online access.
And I am in fact online while running these, so this is all strange.

This part of the tests seems long abandoned.