Remove bin/oltest script.
'oltest' was short for 'offline test'.
This was intended for running only tests that do not need an internet connection.
But when I switch off my WiFi connection, I can run `bin/test` just fine, which now runs all available tests of both `zc.buildout` and `zc.recipe.egg`.

Note: this script had nothing to do with the 'offline' buildout mode/option.  But it may have helped that I removed a few broken tests for that.