Log all http errors when processing package url.
PyPI is having issues right now.  Sometimes I get a timeout, which is reported correctly.
Other times I get an `HTTP Error 503: Service Unavailable`.  This is not reported, so you have no idea of the cause when later on this leads to a DistributionNotFound.

See https://github.com/plone/buildout.coredev/pull/1013#issuecomment-2785929350