Override pkg_resources.Environment.can_add to have better results on Mac.
The override only changes the result on Mac: other platforms seem not affected.

Two minor other changes in here:

* Added some `logger.debug` lines in `_move_to_eggs_dir_and_compile`.
* Added option to use uv for creating the test environment by setting environment variable USE_UV a non-empty value.  We don't use this in our current test setup on GitHub Actions though.

See original report by Jens Vagelpohl a few years ago for setuptools:
https://github.com/pypa/setuptools/issues/3687

And his comment in zopefoundation/meta explains well what happens from
the standpoint of Buildout:
https://github.com/zopefoundation/meta/issues/181#issuecomment-1757558915

Nowadays, the same can happen with a Python installed by uv, as I saw today:

```
% bin/zopepy
>>> import pkg_resources
>>> pkg_resources.get_platform()
'macosx-11.0-arm64'
>>> pkg_resources.get_supported_platform()
'macosx-15.4-arm64'
```

Here `macosx-11.0` is the platform on which the Python was built/compiled.
And `macosx-15.4` is the current platform (my laptop).

This gives problems when we get a Mac-specific wheel.  We turn it into an
egg that has the result of `pkg_resources.get_supported_platform()` in its name.
Then our code in `easy_install._get_matching_dist_in_location` creates a
`pkg_resources.Environment` with the egg location.  Under the hood,
`pkg_resources.compatible_platforms` is called, and this does not find any
matching dists because it compares the platform in the egg name with that
of the system, which is `pkg_resources.get_platform()`.

As far as I see, only Mac has this problem, depending on your Python and
on the packages that you install:

- uv installs a pre-built Python, so the platforms likely differ.
  But this PR should fix that.
- When using pyenv, you compile a Python locally, so both platforms are exactly the same, so there is no problem.
- When you upgrade to a new Mac OSX version, but keep the previously compiled pyenv, the platforms will differ.
  You should reinstall all pyenv Pythons.  But this PR should fix that.
- Wheels with only Python code (no C, Rust, etc), will have a name ending in py3-none-any.whl, and this does not suffer from this problem.