Temporary: test only on MacOS 15.
This will become the default of macos-latest in August:
https://github.com/actions/runner-images/issues/12520
So let's test if that affects us.

This is just a temporary try-out, we don't need to merge this PR.
This PR only runs the Mac tests, no others.
For good measure, I test with a matrix of two Python versions and two setuptools versions.