GHA: do not cache eggs.
We are installing eggs (or at least putting stuff in the eggs directory), and need to see if this still works.
Case in point: I am wondering if we are getting a cached zope.interface egg which still has pkg_resources namespace files.
But: GHA says there is currently no cache.