easy_install.build: don't require presence of a setup.py file.
This is mainly for building eggs with custom options, so quite likely there *is* a setup.py file.
But maybe there is a pyproject.toml file or something else that pip can use.
So when we don't find a setup.py file, we don't raise an exception, but we log a warning.