Develop: use the .pth files that pip creates.
This replaces the .egg-link files.
With setuptools older than 80, .egg-link files may still be created, and then we use them.
Note that this may not be the setuptools version that you have installed next to zc.buildout:
a package that you install may have a pyproject.toml with a build-system that specifies an older setuptools.