Fix failing Windows checks for upgrading.
We get far too many extra lines on Windows:

```
Not upgrading because not running a local buildout command.
```

I found the reason: on Windows sys.argv can have '\\\\?' as prefix...

Another problem.  Now that we include horse-with-no-namespace as dependency, buildout often thinks it needs to be upgraded (buildout restarted).
Reason: we compare two paths, where one has as drive for example `D:` and the other `d:`.  We now compare them with normalized case.

We had excluded the `update.txt` tests on Windows for a long time, because of the above reasons.  Now that these problems have somehow spread to the other tests, I fixed them.  So we now run the `update.txt` tests on Windows as well.